// src/app/page/master/map-fasilitas-kelas/map-fasilitas-kelas.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapFasilitasKelas {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}